Same for rich text pkg
